split init function from workspacebrowser
