+++ rnote-ui/data/ui/penssidebar/penssidebar.ui
