+++ rnote-ui/src/workspacebrowser/components/mod.rs
@@ -1,0 +1,15 @@
+mod remove_workspace_button;
+mod edit_workspace_button;
+mod add_workspace_button;
+mod workspace_list;
+mod workspace_listbox;
+mod files_prefix_listbox;
+mod file_rows;
+
+pub use add_workspace_button::setup_add_workspace_button;
+pub use edit_workspace_button::setup_edit_workspace_button;
+pub use remove_workspace_button::setup_remove_workspace_button;
+pub use workspace_list::setup_workspace_list;
+pub use workspace_listbox::setup_workspace_listbox;
+pub use files_prefix_listbox::setup_files_prefix_listbox;
+pub use file_rows::setup_file_rows;
