+++ rnote-ui/src/workspacebrowser/mod.rs
@@ -6,6 +6,9 @@ mod workspacelist;
 mod workspacelistentry;
 mod workspacerow;
 
+mod components;
+
+use cairo::glib::Cast;
 // Re-exports
 pub use filerow::FileRow;
 pub use workspacelist::WorkspaceList;
@@ -14,11 +17,8 @@ pub use workspacerow::WorkspaceRow;
 
 use crate::appwindow::RnoteAppWindow;
 use gtk4::{
-    gdk, gio, glib, glib::clone, glib::closure, prelude::*, subclass::prelude::*, Button,
-    CompositeTemplate, ConstantExpression, CustomSorter, DirectoryList, FileFilter, FilterChange,
-    FilterListModel, Grid, ListBox, ListBoxRow, ListItem, ListView, MultiSorter,
-    PropertyExpression, ScrolledWindow, SignalListItemFactory, SingleSelection, SortListModel,
-    SorterChange, Widget,
+    gdk, gio, glib, prelude::*, subclass::prelude::*, Button, CompositeTemplate, DirectoryList,
+    Grid, ListBox, ListBoxRow, ListView, ScrolledWindow, Widget,
 };
 use std::path::PathBuf;
 
@@ -163,275 +163,19 @@ impl WorkspaceBrowser {
     }
 
     pub fn init(&self, appwindow: &RnoteAppWindow) {
-        let remove_workspace_button = self.imp().remove_workspace_button.get();
-        let edit_workspace_button = self.imp().edit_workspace_button.get();
+        let remove_workspace_button = components::setup_remove_workspace_button(self, appwindow);
+        let edit_workspace_button = components::setup_edit_workspace_button(self, appwindow);
+        let _add_workspace_button = components::setup_add_workspace_button(self, appwindow);
 
-        self.imp().add_workspace_button.get().connect_clicked(
-            clone!(@weak self as workspacebrowser, @weak appwindow => move |_add_workspace_button| {
-                let dir = workspacebrowser.selected_workspace_dir().unwrap_or(PathBuf::from("./"));
-                workspacebrowser.add_workspace(dir);
-
-                // Popup the edit dialog after creation
-                adw::prelude::ActionGroupExt::activate_action(&appwindow, "edit-workspace", None);
-            }),
+        components::setup_workspace_list(
+            self,
+            appwindow,
+            &remove_workspace_button,
+            &edit_workspace_button,
         );
-
-        self.imp().remove_workspace_button.get().connect_clicked(
-            clone!(@weak self as workspacebrowser, @weak appwindow => move |_| {
-                workspacebrowser.remove_current_workspace();
-            }),
-        );
-
-        self.imp().edit_workspace_button.get().connect_clicked(
-            clone!(@weak appwindow => move |_| {
-                adw::prelude::ActionGroupExt::activate_action(&appwindow, "edit-workspace", None);
-            }),
-        );
-
-        self.imp().workspace_list.connect_items_changed(
-            clone!(@weak self as workspacebrowser, @weak appwindow, @weak remove_workspace_button, @weak edit_workspace_button => move |folders_model, _, _, _| {
-                remove_workspace_button.set_sensitive(folders_model.n_items() > 1);
-                edit_workspace_button.set_sensitive(folders_model.n_items() > 0);
-                workspacebrowser.save_to_settings(&appwindow.app_settings());
-            }),
-        );
-
-        self.imp().workspace_listbox.connect_selected_rows_changed(clone!(@weak appwindow, @weak self as workspacebrowser => move |_| {
-            if let Some(dir) = workspacebrowser.current_selected_workspace_row().map(|row| row.entry().dir()) {
-                workspacebrowser.imp().files_dirlist.set_file(Some(&gio::File::for_path(dir)));
-                workspacebrowser.save_to_settings(&appwindow.app_settings());
-            }
-
-        }));
-
-        // Setup prefix listbox
-        self.imp().files_prefix_listbox.connect_row_activated(clone!(@weak self as workspacebrowser, @weak appwindow => move |_, row| {
-            if row == &workspacebrowser.imp().dir_up_row.get() {
-                if let Some(parent_dir) = workspacebrowser.selected_workspace_dir().and_then(|p| p.parent().map(|p| p.to_path_buf())) {
-                    workspacebrowser.set_current_workspace_dir(parent_dir.to_path_buf());
-                }
-
-            }
-        }));
-
-        // Setup file rows
-        {
-            let primary_list_factory = SignalListItemFactory::new();
-
-            primary_list_factory.connect_setup(clone!(@weak appwindow => move |_, list_item| {
-                let filerow = FileRow::new();
-                filerow.init(&appwindow);
-
-                list_item.set_child(Some(&filerow));
-
-                let list_item_expr = ConstantExpression::new(list_item);
-                let fileinfo_expr =
-                    PropertyExpression::new(ListItem::static_type(), Some(&list_item_expr), "item");
-
-                let file_expr = fileinfo_expr.chain_closure::<Option<gio::File>>(closure!(
-                    |_: Option<glib::Object>, fileinfo_obj: Option<glib::Object>| {
-                        fileinfo_obj
-                            .map(|fileinfo_obj| {
-                                fileinfo_obj
-                                    .downcast::<gio::FileInfo>()
-                                    .unwrap()
-                                    .attribute_object("standard::file")
-                                    .unwrap()
-                                    .downcast::<gio::File>()
-                                    .unwrap()
-                            })
-                            .to_value()
-                    }
-                ));
-
-                let content_provider_expr =
-                    fileinfo_expr.chain_closure::<gdk::ContentProvider>(closure!(
-                        |_: Option<glib::Object>, fileinfo_obj: Option<glib::Object>| {
-                            if let Some(fileinfo_obj) = fileinfo_obj {
-                                if let Some(file) = fileinfo_obj
-                                    .downcast::<gio::FileInfo>()
-                                    .unwrap()
-                                    .attribute_object("standard::file")
-                                {
-                                    let file = file
-                                        .downcast::<gio::File>()
-                                        .expect("failed to downcast::<gio::File>() from file GObject");
-
-                                    return gdk::ContentProvider::for_value(&file.to_value());
-                                }
-                            }
-
-                            gdk::ContentProvider::for_value(&None::<gio::File>.to_value())
-                        }
-                    ));
-
-                let icon_name_expr =
-                    fileinfo_expr.chain_closure::<gio::ThemedIcon>(closure!(|_: Option<glib::Object>, fileinfo_obj: Option<glib::Object>| {
-                        if let Some(fileinfo_obj) = fileinfo_obj {
-                            if let Some(themed_icon) = fileinfo_obj
-                                .downcast::<gio::FileInfo>()
-                                .unwrap()
-                                .attribute_object("standard::icon")
-                            {
-                                return themed_icon.downcast::<gio::ThemedIcon>().unwrap();
-                            }
-                        }
-
-                        gio::ThemedIcon::from_names(&[
-                            "workspace-folder-symbolic",
-                            "folder-documents-symbolic",
-                        ])
-                    }));
-
-                let basename_expr =
-                    fileinfo_expr.chain_closure::<String>(closure!(|_: Option<glib::Object>, fileinfo_obj: Option<glib::Object>| {
-                        if let Some(fileinfo_obj) = fileinfo_obj {
-                            if let Some(file) = fileinfo_obj
-                                .downcast::<gio::FileInfo>()
-                                .unwrap()
-                                .attribute_object("standard::file")
-                            {
-                                let file = file
-                                    .downcast::<gio::File>()
-                                    .expect("failed to downcast::<gio::File>() from file GObject");
-
-                                return String::from(
-                                    file.basename()
-                                        .expect("failed to get file.basename()")
-                                        .to_string_lossy(),
-                                );
-                            }
-                        }
-
-                        String::from("")
-                    }));
-
-                file_expr.bind(&filerow, "current-file", Widget::NONE);
-                basename_expr.bind(&filerow.file_label(), "label", Widget::NONE);
-                icon_name_expr.bind(&filerow.file_image(), "gicon", Widget::NONE);
-                content_provider_expr.bind(&filerow.drag_source(), "content", Widget::NONE);
-            }));
-
-            let filefilter = FileFilter::new();
-            filefilter.add_pattern("*.rnote");
-            filefilter.add_pattern("*.xopp");
-            filefilter.add_pattern("*.svg");
-            filefilter.add_mime_type("image/svg+xml");
-            filefilter.add_mime_type("image/png");
-            filefilter.add_mime_type("image/jpeg");
-            filefilter.add_mime_type("application/x-xopp");
-            filefilter.add_mime_type("application/pdf");
-            filefilter.add_mime_type("inode/directory");
-            let filefilter_model =
-                FilterListModel::new(Some(&self.imp().files_dirlist), Some(&filefilter));
-
-            let folder_sorter = CustomSorter::new(move |obj1, obj2| {
-                let first_fileinfo = obj1
-                    .clone()
-                    .downcast::<gio::FileInfo>()
-                    .expect("failed to downcast obj1");
-                let first_filetype = first_fileinfo.file_type();
-
-                let second_fileinfo = obj2
-                    .clone()
-                    .downcast::<gio::FileInfo>()
-                    .expect("failed to downcast obj2");
-                let second_filetype = second_fileinfo.file_type();
-
-                if first_filetype == gio::FileType::Directory
-                    && second_filetype != gio::FileType::Directory
-                {
-                    gtk4::Ordering::Smaller
-                } else if first_filetype != gio::FileType::Directory
-                    && second_filetype == gio::FileType::Directory
-                {
-                    gtk4::Ordering::Larger
-                } else {
-                    gtk4::Ordering::Equal
-                }
-            });
-
-            let alphanumeric_sorter = CustomSorter::new(move |obj1, obj2| {
-                let first_fileinfo = obj1
-                    .clone()
-                    .downcast::<gio::FileInfo>()
-                    .expect("failed to downcast obj1");
-                let first_file = first_fileinfo.attribute_object("standard::file").unwrap();
-                let first_file = first_file.downcast::<gio::File>().unwrap();
-                let first_display_name = first_file.basename().unwrap();
-                let first_display_name = first_display_name.to_str().unwrap();
-
-                let second_fileinfo = obj2
-                    .clone()
-                    .downcast::<gio::FileInfo>()
-                    .expect("failed to downcast obj2");
-                let second_file = second_fileinfo.attribute_object("standard::file").unwrap();
-                let second_file = second_file.downcast::<gio::File>().unwrap();
-                let second_display_name = second_file.basename().unwrap();
-                let second_display_name = second_display_name.to_str().unwrap();
-
-                first_display_name.cmp(second_display_name).into()
-            });
-
-            let multisorter = MultiSorter::new();
-            multisorter.append(&folder_sorter);
-            multisorter.append(&alphanumeric_sorter);
-            let multi_sort_model = SortListModel::new(Some(&filefilter_model), Some(&multisorter));
-
-            let primary_selection_model = SingleSelection::new(Some(&multi_sort_model));
-
-            self.imp()
-                .files_listview
-                .get()
-                .set_factory(Some(&primary_list_factory));
-            self.imp()
-                .files_listview
-                .get()
-                .set_model(Some(&primary_selection_model));
-
-            self.imp().files_listview.get().connect_activate(clone!(@weak filefilter, @weak multisorter, @weak appwindow => move |files_listview, position| {
-                let model = files_listview.model().expect("model for primary_listview does not exist.");
-                let fileinfo = model.item(position)
-                    .expect("selected item in primary_listview does not exist.")
-                    .downcast::<gio::FileInfo>().expect("selected item in primary_list is not of Type `gio::FileInfo`");
-
-                if let Some(file) = fileinfo.attribute_object("standard::file") {
-                    let file = file.downcast::<gio::File>().unwrap();
-
-                    appwindow.open_file_w_dialogs(&file, None);
-                };
-
-                multisorter.changed(SorterChange::Different);
-                filefilter.changed(FilterChange::Different);
-            }));
-
-            self.imp().files_dirlist.connect_file_notify(
-                clone!(@weak self as workspacebrowser, @weak appwindow, @weak filefilter, @weak multisorter => move |files_dirlist| {
-                    // Disable the dir up row when no file is set or has no parent
-                    workspacebrowser.imp().dir_up_row.set_sensitive(files_dirlist.file().and_then(|f| f.parent()).is_some());
-
-                    multisorter.changed(SorterChange::Different);
-                    filefilter.changed(FilterChange::Different);
-                }),
-            );
-
-            self.imp().files_dirlist.connect_items_changed(clone!(@weak filefilter, @weak multisorter => move |_primary_dirlist, _position, _removed, _added| {
-                multisorter.changed(SorterChange::Different);
-                filefilter.changed(FilterChange::Different);
-            }));
-
-            // setup workspace rows
-            let appwindow_c = appwindow.clone();
-            self.imp()
-                .workspace_listbox
-                .bind_model(Some(&self.imp().workspace_list), move |obj| {
-                    let entry = obj.to_owned().downcast::<WorkspaceListEntry>().unwrap();
-                    let workspace_row = WorkspaceRow::new(entry);
-                    workspace_row.init(&appwindow_c);
-
-                    workspace_row.upcast::<Widget>()
-                });
-        }
+        components::setup_workspace_listbox(self, appwindow);
+        components::setup_files_prefix_listbox(self, appwindow);
+        components::setup_file_rows(self, appwindow);
 
         self.setup_dir_actions(appwindow);
     }
@@ -553,6 +297,8 @@ impl WorkspaceBrowser {
     fn setup_dir_actions(&self, _appwindow: &RnoteAppWindow) {
         self.insert_action_group("workspace_action", Some(&self.imp().workspace_actions));
 
-        self.imp().workspace_actions.add_action(&workspace_action::create_dir(self));
+        self.imp()
+            .workspace_actions
+            .add_action(&workspace_action::create_dir(self));
     }
 }
